Incorrect Space Complexity
Hi,
I tired calculating the space taken  by the stack in the program above and worst case  when you have a tree like
           1
       /       \
     2          3 
   /   \
 4      5

The height of the tree is two but worst case space consumed is 5. This can be extended to bigger trees but the space taken is much bigger than the height of the tree and the next logical space bound is n. Please let me know if you need more details.